rip out more types
